==> code/02-data-preprocessing.ipynb <==
# Data Pre-Processing

First, we are going to clean the data and then we are going to perform feature engineering to enrich the data.


## 1. Data Cleaning

First, we are going to clean the data by:
- checking for duplicate rows
- checking for missing values
- checking the percentage of missing values of each column
- checking for any redundant (uninformative) columns
- checking the column data types

import os
import re
import ast
import pandas as pd
import numpy as np
import spacy         # Natural language processing
import isodate       # Date transformation and manipulation

from os.path import dirname, abspath
from dateutil import parser
from datetime import datetime, date

#### Read the channels and videos files into dataframes

# Read in the csv files from the raw data folder
project_dir = dirname(dirname(abspath("02-data-preprocessing.ipynb")))
channels_df = pd.read_csv(project_dir + "/data/raw/fitness_channels_2023_07_11.csv")
videos_df = pd.read_csv(project_dir + "/data/raw/fitness_videos_2023_07_11.csv")

# rename the columns to make the naming more intuitive and consistent
channels_df = channels_df.rename(columns={
    'ChannelName': 'channel',
    'ChannelDescription': 'channel_description',
    'PublishedDate': 'published_date',
    'TotalSubscribers': 'total_subscribers',
    'TotalViews':'total_views',
    'TotalVideos': 'total_videos', 
    'playlistID': 'playlist_id'
    
})

videos_df = videos_df.rename(columns={
    'channelTitle': 'channel',
    'viewCount': 'total_views',
    'likeCount': 'total_likes',
    'commentCount': 'total_comments',
    'favouriteCount': 'total_favourites',
    'publishedAt': 'published_at'
})

#### Check for duplicate rows and columns

# check for duplicate rows in channels and videos data
(channels_df.duplicated().any(),videos_df.duplicated().any())

#### Check for missing data

videos_df.isnull().any()

#### Check for percentage of missing values

videos_df.isnull().sum() / videos_df.shape[0] * 100.00

# Drop rows with missing values for likes - those videos are shorts
videos_df = videos_df.dropna(subset=['total_likes'])

# Drop the favouriteCount column since it only contains missing values
videos_df = videos_df.drop(columns='total_favourites')

# Fill missing comments with 0 - comments was not allowed for those videos
videos_df['total_comments'] = videos_df['total_comments'].fillna(0)


#### Check for redundant columns

# Check the percentage of each unique value in the definition column
videos_df['definition'].value_counts() / videos_df['definition'].shape[0] * 100

# Drop the definition column for being uninformative
videos_df = videos_df.drop(columns='definition')

# Check the date values to make sure there are no errors
videos_df.published_at.sort_values()

#### Checking the column data types

videos_df.dtypes

## 2. Feature Engineering

Second, we are going to perform feature engineering to enrich the data by:
* Convert ISO 8601 duration format from the YouTube Data API v3
    1. First convert to timedelta[s]
    2. Convert duration from units of seconds to minutes since it is more intuitive for workout videos
    3. Check for any errors or outliers (address them if necessary)
    4. Categorize the duration columns into groups
* Derive new features from the published date column 
    1. Get published year
    2. Get published month
    3. Get published day of the week
    4. Get the age of the video
* Get the length of the video title
* Get the total number of tags in the tags column
* use natural language processing (via spaCy) to derive useful features from the video title, description, and/or tags
    1. Get the workout length; since actual workout duration and video duration are usually not the same
    2. Get the workout type (cardio, HIIT, yoga, pilates, etc..)
    3. Get the target body part (legs, abs, arms ,etc..)
    3. Get any special aspect of the workout (no jumping, standing, no equipment, etc..)

# Convert duration to seconds
videos_df['duration_secs'] = videos_df['duration'].apply(lambda x: isodate.parse_duration(x))
videos_df['duration_mins'] = videos_df['duration_secs'].astype('timedelta64[s]')/60

# Check the distribution of the video duration for presence of any outliers
videos_df['duration_mins'].describe()

# Dropping videos that have 0 mins as duration - most likely livestreams that might have accidentally got turned on and off right away
videos_df = videos_df[videos_df['duration_mins'] > 0]

# Check for videos that are extremely long, greater than 45 mins.
# Note: we get that there are 20 videos that are that long
videos_df[videos_df['duration_mins'] > 45][['channel','title','duration','video_id']]

# Define the bin edges for the duration ranges
bin_edges = [0, 5, 10, 20, 30, 45, float('inf')]

# Define the labels for each range
labels = ['0-5', '5-10', '10-20', '20-30', '30-45', '45+']

# Use pd.cut to categorize the 'duration_mins' column into the specified ranges
videos_df['duration_category'] = pd.cut(videos_df['duration_mins'], bins=bin_edges, labels=labels, right=False)

# Create publish year and month (of the year) columns
videos_df['published_datetime'] = pd.to_datetime(videos_df['published_at'].apply(lambda x: parser.parse(x)))
videos_df['published_date'] = pd.to_datetime(videos_df['published_datetime'].dt.date)

# Extract the month and year from the 'published_date' to create new columns
videos_df['published_month'] = videos_df['published_date'].dt.month
videos_df['published_year'] = videos_df['published_date'].dt.year

# Get the current date 
current_date_str = date.today()
current_date = pd.to_datetime(current_date_str)

# Calculate the video age in months
videos_df['video_age'] = ((current_date.year - videos_df['published_year']) * 12 +
                         (current_date.month - videos_df['published_month']))

# Convert the description column to string
videos_df['description'] = videos_df['description'].astype(str)

# Title character length
videos_df['title_length'] = videos_df['title'].str.len()

# Create the number of tags column
videos_df['tags'] = videos_df['tags'].replace(np.nan, None)
videos_df['tags'] = videos_df['tags'].apply(lambda x: x if x is None else ast.literal_eval(x))
videos_df['total_tags'] = videos_df['tags'].apply(lambda x: 0 if x is None else len(x))

# Now we are going to use spaCy's entity ruler along with regex 
# to look for specific entities in the text columns.

# Download spaCy's small english model
nlp = spacy.load("en_core_web_sm")

# Create and add the EntityRuler
ruler = nlp.add_pipe("entity_ruler", before="ner")

#List of Entities and Patterns
patterns = [
    # labels for body part target
    {"label": "FULL_BODY", "pattern": [{"LOWER": {"REGEX": r"(full|total|whole)"}}, {"LOWER": "body"}]},
    {"label": "UPPER_BODY", "pattern": [{"LOWER": "upper"}, {"LOWER": "body"}]},
    {"label": "LOWER_BODY", "pattern": [{"LOWER": "lower"}, {"LOWER": "body"}]},
    {"label": "CHEST", "pattern": [{"LOWER": "chest"}]},
    {"label": "BACK", "pattern": [{"LOWER": "back"}]},
    {"label": "ABS", "pattern": [{"LOWER": {"REGEX": r"(core|ab|abs|plank)"}}]},
    {"label": "ARMS", "pattern": [{"LOWER": {"REGEX": r"arms?"}}]},
    {"label": "LEGS", "pattern": [{"LOWER": {"REGEX": r"(thigh|thighs|leg|legs)"}}]},
    {"label": "GLUTES", "pattern": [{"LOWER": {"REGEX": r"(booty|glute|glutes|butt)"}}]},
    # labels for workout type
    {"label": "HIIT", "pattern": [{"LOWER": "hiit"}]},
    {"label": "CARDIO", "pattern": [{"LOWER": "cardio"}]},
    {"label": "DANCE", "pattern": [{"LOWER": "dance"}]},
    {"label": "TABATA", "pattern": [{"LOWER": "tabata"}]},
    {"label": "PILATES", "pattern": [{"LOWER": "pilates"}]},
    {"label": "BARRE", "pattern": [{"LOWER": "barre"}]},
    {"label": "YOGA", "pattern": [{"LOWER": "yoga"}]},
    {"label": "STANDING", "pattern": [{"LOWER": "standing"}]},
    {"label": "NO_EQUIPMENT", "pattern": [{"LOWER": "no", "LOWER": {"REGEX": r"(equip|equipment|equipments|weight|weights)"}}]},
    {"label": "NO_JUMPING", "pattern": [{"LOWER": "no", "LOWER": "jumping"}]},
    {"label": "LOW_IMPACT", "pattern": [{"LOWER": "low", "LOWER": "impact"}]},
    {"label": "STRENGTH_TRAINING", "pattern": [{"LOWER": {"REGEX": r"(strength|sculpt|sculpting|tone|toning|toned)"}}]}
]

ruler.add_patterns(patterns)


def extract_ent_label(string, label):
    """ Identify presence of label using regular expressions

    Arguments:
        string -- the string value of either the title, description columns
        label -- can be either one of the labels mentioned in the patterns list above

    Returns:
        A logical value, True if the label is found, else the default value is False
    """
    doc = nlp(string)
    result = False
    for ent in doc.ents:
        if ent.label_ == label:
            result = True
            break
    return result


# Extract the values for the "label" key and store them as a list
labels_list = [pattern['label'] for pattern in patterns]

# Create new columns to the exting videos dataframe
videos_df = videos_df.assign(**{label: None for label in labels_list})

# Apply the extract_ent_label function to the new columns row-wise using apply and lambda function
for label in labels_list:
    print("Getting video text information for column: " + label)
    
    # extract text information about the workout from the video title first
    videos_df[label] = videos_df.apply(lambda row: extract_ent_label(row['title'], label), axis=1)

    # if missing from the title, extract text information about the workout from the video description
    videos_df[label] = videos_df.apply(lambda row: extract_ent_label(row['description'], label)\
        if (not row[label]) else row[label], axis=1)

videos_df.info()

# Drop the columns we no longer need
videos_df = videos_df.drop(columns=['duration','published_at','duration_secs'])

#### Save the processed data 

# Save dataframes as csv files in the processed subfolder of the data folder
channels_df.to_csv(project_dir + "/data/processed/fitness_channels_processed_2023_07_11.csv", index=False)
videos_df.to_csv(project_dir + "/data/processed/fitness_videos_processed_2023_07_11.csv", index=False)
