==> code/02_data_preprocessing.ipynb <==
## Data Pre-Processing

import os
import ast
import isodate
import pandas as pd
import numpy as np

from dateutil import parser

#### Read the channels and videos files into dataframes

cwd = os.path.dirname(os.getcwd())
path = cwd + "/data/raw"
channels_df = pd.read_csv(path + "/channels_data.csv")
videos_df = pd.read_csv(path + "/videos_data.csv")

channels_df.head()

videos_df.head()

#### Drop redundant column and rows

channels_df.drop('Unnamed: 0', axis = 1, inplace = True)
videos_df.drop('Unnamed: 0', axis = 1, inplace = True)

(channels_df.duplicated().any(),videos_df.duplicated().any())

#### Missing data

channels_df.isnull().any()

Getting video information from channel: blogilates
Could not get comments for video AiVqvB07GyI
Could not get comments for video nPQHHZ5sjTw
Could not get comments for video X1kMiRQq354
Could not get comments for video k9mZ6zc6wM0
Could not get comments for video o3E9Oak_n7E
Could not get comments for video kCWmpm5nyCo
Could not get comments for video pPHl8qfAo2A
Could not get comments for video Tfei0NvjZrM

Getting video information from channel: Rebecca-Louise
Could not get comments for video SB8zu5kprJY
Could not get comments for video oAax6YZvaN0
Could not get comments for video 6GsbnH-Tw7k
Could not get comments for video 3vJ05j99CT8
Could not get comments for video rbxxmB6cqtE
Could not get comments for video b7PZXm4NKqk
Could not get comments for video BDDZ9X8-rYg
Could not get comments for video wbH0eD3b7D4

Getting video information from channel: emi wong
Could not get comments for video NJZ95sZlXL4

Getting video information from channel: MadFit
Could not get comments for video VJ75BRoB24I

videos_df.isnull().any()

# Find the percentage of missing values from columns that contains them
missingval_columns = videos_df.loc[:, ['description', 'tags', 'likeCount','favouriteCount','commentCount']]
missingval_columns.isnull().sum() / missingval_columns.shape[0] * 100.00

videos_df.drop('favouriteCount', axis=1, inplace=True)

videos_df['definition'].unique()

videos_df['definition'].value_counts() / videos_df['definition'].shape[0] * 100

videos_df.drop('definition', axis=1, inplace=True)

videos_df.publishedAt.sort_values()

#### Checking data types

channels_df.dtypes

videos_df.dtypes

## Feature Engineering

# Title character length
videos_df['titleLength'] = videos_df['title'].apply(lambda x: len(x))

# Convert duration to seconds
videos_df['durationSecs'] = videos_df['duration'].apply(lambda x: isodate.parse_duration(x))
videos_df['durationSecs'] = videos_df['durationSecs'].astype('timedelta64[s]')

# Create publish year and month (of the year) columns
videos_df['publishedDatetime'] =  videos_df['publishedAt'].apply(lambda x: parser.parse(x))
videos_df['publishedYear'] = videos_df['publishedDatetime'].apply(lambda x: x.strftime("%Y"))
videos_df['publishedMonth'] = videos_df['publishedDatetime'].apply(lambda x: x.strftime("%b"))

# Create publish day (of the week) and hour and time of day columns
videos_df['pushblishDayName'] = videos_df['publishedDatetime'].apply(lambda x: x.strftime("%a")) 
videos_df['publishedHour'] = videos_df['publishedDatetime'].apply(lambda x: x.strftime("%H"))
videos_df['publishedToD'] = videos_df['publishedDatetime'].apply(lambda x: x.strftime("%p"))


# Create the number of tags column
videos_df['tags'] = videos_df['tags'].replace(np.nan, None)
videos_df['tags'] = videos_df['tags'].apply(lambda x: x if x is None else ast.literal_eval(x))
videos_df['tagsCount'] = videos_df['tags'].apply(lambda x: 0 if x is None else len(x))

# Create column to categorize Youtube Shorts
videos_df['youtubeShorts'] = [True if x <= 60 else False for x in videos_df['durationSecs']]

videos_df.dtypes

videos_df.head()

Natural Language Processing of text (title column) to enrich the data

workout_videos_df = videos_df[videos_df['youtubeShorts'] == False].copy()

workout_videos_df.shape

import spacy
nlp = spacy.load("en_core_web_sm")

#Create the EntityRuler
ruler = nlp.add_pipe("entity_ruler", before="ner")

#List of Entities and Patterns
patterns = [
    {"label": "WORKOUT_TIME", "pattern": [{"TEXT": {"REGEX": r"^(\d+)"}},
                                          {"LOWER":{"REGEX": r"^(min|mins|minute|minutes|hour|hours|hr|hrs)$"}}
                                         ]},
    {"label": "FULL_BODY", "pattern": [{"LOWER": {"REGEX": r"(full|total|whole)"}}, {"LOWER": "body"}]},
    {"label": "UPPER_BODY", "pattern": [{"LOWER": "upper"}, {"LOWER": "body"}]},
    {"label": "LOWER_BODY", "pattern": [{"LOWER": "lower"}, {"LOWER": "body"}]},
    {"label": "CHEST_BACK", "pattern": [{"LOWER": {"REGEX": r"(back|chest)"}}, {"ORTH": {"REGEX": r"(and|&)"}} ,{"LOWER": {"REGEX": r"(back|chest)"}}]},
    {"label": "ABS", "pattern": [{"LOWER": {"REGEX": r"(core|ab|abs)"}}]},
    {"label": "ARMS", "pattern": [{"LOWER": {"REGEX": r"arms?"}}]},
    {"label": "LEGS", "pattern": [{"LOWER": {"REGEX": r"(thigh|thighs|leg|legs)"}}]},
    {"label": "GLUTES", "pattern": [{"LOWER": {"REGEX": r"(booty|glute|glutes|butt)"}}]}
]

ruler.add_patterns(patterns)

# Extract workout time using regular expressions
def extract_workout_time(title):
    doc = nlp(title)
    workout_time = None
    for ent in doc.ents:
        if ent.label_ == "WORKOUT_TIME":
            workout_time = ent.text
            break
    return workout_time


# Apply the function to create a workout time column
workout_videos_df['workoutTime'] = workout_videos_df['title'].apply(extract_workout_time)

# Function to extract body part from a title
def extract_body_part(title):
    doc = nlp(title)
    body_part = None
    body_parts = ["FULL_BODY","UPPER_BODY","LOWER_BODY","CHEST_BACK","ARMS","ABS","LEGS","GLUTES"]
    for ent in doc.ents:
        if ent.label_ in body_parts:
            body_part = (ent.text, ent.label_)
            break
    return body_part


# Apply the function to create a body part column
workout_videos_df['bodyPart'] = workout_videos_df['title'].apply(extract_body_part)

workout_videos_df.head()
